Migrate IOThread tests
